move Hero to global components, move hero styles to seperate file, add styles for each page variation, update imports
